+++ src/test/java/hw3/TestFourTextImages.java
@@ -9,10 +9,14 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.Properties;
+
 public class TestFourTextImages {
     private WebDriver driver;
     FourTextImages objFourTextImages;
-
+    Properties properties = loadProperties();
 
     @DataProvider(name="urlAndTexts")
     public Object[][] dpMethod(){
@@ -31,9 +35,17 @@ public class TestFourTextImages {
     public void setupTest() {
         driver = new ChromeDriver();
         objFourTextImages = new FourTextImages(driver);
-        driver.navigate().to("https://jdi-testing.github.io/jdi-light/index.html");
+        driver.navigate().to(properties.getProperty("page.homeUrl"));
     }
-
+    private Properties loadProperties() {
+        try (InputStream input = TestMenuItems.class.getClassLoader().getResourceAsStream("resources.properties")) {
+            Properties prop = new Properties();
+            prop.load(input);
+            return prop;
+        } catch (IOException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
     @Test(dataProvider = "urlAndTexts")
     public void isTextCorrect(WebElement webElem, String menuTextItem) {
         SoftAssert isTextCorrect = new SoftAssert();
