+++ src/test/java/hw3/TestFrames.java
@@ -5,14 +5,19 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 import org.testng.asserts.SoftAssert;
 
+import java.io.IOException;
+import java.io.InputStream;
 import java.util.List;
+import java.util.Properties;
 
 public class TestFrames {
     private WebDriver driver;
+    Properties properties = loadProperties();
 
     @BeforeClass
     public static void setupClass() { WebDriverManager.chromedriver().setup();}
@@ -20,9 +25,19 @@ public class TestFrames {
     @BeforeClass
     public void setupTest() {
         driver = new ChromeDriver();
-        driver.navigate().to("https://jdi-testing.github.io/jdi-light/index.html");
+        driver.navigate().to(properties.getProperty("page.homeUrl"));
     }
+    @AfterClass public void closeBrowser() {driver.close();}
 
+    private Properties loadProperties() {
+        try (InputStream input = TestMenuItems.class.getClassLoader().getResourceAsStream("resources.properties")) {
+            Properties prop = new Properties();
+            prop.load(input);
+            return prop;
+        } catch (IOException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
     @Test
     public void isButtonInFrameExist () {
         SoftAssert buttonInFrameExist = new SoftAssert();
